Get now playing drawer to act nicely with navigation set up
