+++ src/App.tsx
@@ -1,15 +1,6 @@
-import React, {
-  useRef,
-  useEffect
-  // useCallback,
-  // useState
-} from 'react'
+import React, { useRef, useEffect } from 'react'
 
-import { PortalProvider } from '@gorhom/portal'
-import {
-  // Animated,
-  Platform
-} from 'react-native'
+import { Platform } from 'react-native'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
 import WebView from 'react-native-webview'
 import { Provider } from 'react-redux'
@@ -24,7 +15,6 @@ import HCaptcha from 'app/components/hcaptcha'
 import MobileUploadDrawer from 'app/components/mobile-upload-drawer'
 import NavigationContainer from 'app/components/navigation-container'
 import Notifications from 'app/components/notifications/Notifications'
-// import NowPlayingDrawer from 'app/components/now-playing-drawer/NowPlayingDrawer'
 import OAuth from 'app/components/oauth/OAuth'
 import OverflowMenuDrawer from 'app/components/overflow-menu-drawer'
 import Search from 'app/components/search/Search'
@@ -92,58 +82,30 @@ const App = () => {
     setupAnalytics()
   }, [])
 
-  /**
-  // Set handlers for the NowPlayingDrawer and BottomTabBar
-  // When the drawer is open, the bottom bar should hide (animated away).
-  // When the drawer is closed, the bottom bar should reappear (animated in).
-  const bottomBarTranslationAnim = useRef(new Animated.Value(0)).current
-  // Track bottom bar display properties as an object, so every update
-  // can be listened to, even if we go from hidden => hidden
-  const [bottomBarDisplay, setBottomBarDisplay] = useState({
-    isShowing: true
-  })
-  const onNowPlayingDrawerOpen = useCallback(() => {
-    setBottomBarDisplay({ isShowing: false })
-  }, [setBottomBarDisplay])
-  const onNowPlayingDrawerClose = useCallback(() => {
-    setBottomBarDisplay({ isShowing: true })
-  }, [setBottomBarDisplay])
-   */
-
   return (
     <SafeAreaProvider>
       <NavigationContainer>
         <Provider store={store}>
-          <PortalProvider>
-            <WebRefContextProvider>
-              <GoogleCast webRef={webRef} />
-              <WebApp webRef={webRef} />
-              {/*
+          <WebRefContextProvider>
+            <GoogleCast webRef={webRef} />
+            <WebApp webRef={webRef} />
+            {/*
                 Note: it is very important that components are rendered after WebApp.
                 On Android, regardless of position: absolute, WebApp will steal all of
                 touch targets and onPress will not work.
               */}
-              <AppNavigator />
-              <Search />
-              <Notifications webRef={webRef} />
-
-              {/*
+            <AppNavigator />
+            {/*
                 Commenting out NowPlayingDrawer until all drawers and overlays are migrated to RN
               */}
-              {/* <NowPlayingDrawer
-                onOpen={onNowPlayingDrawerOpen}
-                onClose={onNowPlayingDrawerClose}
-                bottomBarTranslationAnim={bottomBarTranslationAnim}
-              />
-              */}
-
-              <Drawers />
-              <Modals />
-              <Audio webRef={webRef} />
-              <OAuth webRef={webRef} />
-              <Airplay webRef={webRef} />
-            </WebRefContextProvider>
-          </PortalProvider>
+            <Search />
+            <Notifications webRef={webRef} />
+            <Drawers />
+            <Modals />
+            <Audio webRef={webRef} />
+            <OAuth webRef={webRef} />
+            <Airplay webRef={webRef} />
+          </WebRefContextProvider>
         </Provider>
       </NavigationContainer>
     </SafeAreaProvider>
